Rename Store > NoSqlStore
