+++ nosqlstore-core/src/main/java/net/spals/appbuilder/nosqlstore/core/model/NoSqlStoreOperator.java
@@ -1,0 +1,19 @@
+package net.spals.appbuilder.nosqlstore.core.model;
+
+/**
+ * @author tkral
+ */
+public interface NoSqlStoreOperator {
+
+    enum Standard implements NoSqlStoreOperator {
+        ALL,
+        BETWEEN,
+        EQUAL_TO,
+        GREATER_THAN,
+        GREATER_THAN_OR_EQUAL_TO,
+        LESS_THAN,
+        LESS_THAN_OR_EQUAL_TO,
+        NONE,
+        ;
+    }
+}
