--- store-core/src/main/java/net/spals/appbuilder/store/core/model/StoreRangeKey.java
@@ -1,11 +1,0 @@
-package net.spals.appbuilder.store.core.model;
-
-/**
- * @author tkral
- */
-public interface StoreRangeKey<C extends Comparable<C>> {
-
-   StoreOperator getOperator();
-
-   C getValue();
-}
